feat: DeleteUser method to user_repository.go. Additional consideration may be required such as 'deleted' field on user rather than removing the record outright.
refactor: refreshToken is now sent in the body rather than a cookie. This represents a current sentiment to pivot away from access + refresh based auth to a simpler session based system. Further analysis will determine whether this plan will be enacted

docs: added comment in compression.go to reflect the purpose of an if statement